rvv: fixx vnclip corner

+++ riscv/insns/vnclip_wi.h
@@ -4,7 +4,7 @@ int64_t int_max = INT64_MAX >> (64 - P.VU.vsew);
 int64_t int_min = INT64_MIN >> (64 - P.VU.vsew);
 VI_VVXI_LOOP_NARROW
 ({
-  int64_t result = vs2;
+  int128_t result = vs2;
   unsigned shift = zimm5 & ((sew * 2) - 1);
 
   // rounding
